字典缓存

+++ basis-apps/basis-project-sys/sys-web/src/main/java/com/wudgaby/platform/sys/dict/DictDTO.java
@@ -13,11 +13,11 @@ public class DictDTO {
     /**
      * 字典id
      */
-    private Integer id;
+    private Long id;
     /**
      * 字典父id
      */
-    private Integer pid;
+    private Long pid;
     /**
      * 字典标签
      */
